+++ app/models/feature.rb
@@ -1,6 +1,8 @@
 class Feature
   def self.enabled?(feature_name)
     case feature_name.to_sym
+    when :badge_filter
+      !Rails.env.production?
     when :paywalled_search_results
       true
     when :redirect_db_id_profiles