+++ test/queries/developer_query_test.rb
@@ -130,6 +130,28 @@ class DeveloperQueryTest < ActiveSupport::TestCase
     refute_includes records, blank
   end
 
+  test "filtering by source contributor badge" do
+    source_contributor_developer = create_developer(source_contributor: true)
+    not_source_contributor_developer = create_developer(source_contributor: false)
+    blank = create_developer
+
+    records = DeveloperQuery.new(badges: ["source_contributor"]).records
+
+    assert_includes records, source_contributor_developer
+    refute_includes records, not_source_contributor_developer
+    refute_includes records, blank
+  end
+
+  test "filtering by recently active badge" do
+    recently_active_developer = create_developer
+    not_recently_active_developer = create_developer
+    not_recently_active_developer.update!(updated_at: 2.weeks.ago)
+
+    records = DeveloperQuery.new(badges: ["recently_active"]).records
+    assert_includes records, recently_active_developer
+    refute_includes records, not_recently_active_developer
+  end
+
   test "filtering developers by their bio or hero does not includes all if business has an active subscription" do
     subscribed_business = users(:subscribed_business)
     loves_oss = create_developer(hero: "I love OSS!")
@@ -156,7 +178,8 @@ class DeveloperQueryTest < ActiveSupport::TestCase
       role_levels: [:junior],
       include_not_interested: true,
       search_query: "rails engineer",
-      countries: ["United States"]
+      countries: ["United States"],
+      badges: [:recently_active]
     }
     assert_equal DeveloperQuery.new(filters.dup).filters, filters
   end